requirements.txt
absl-py==2.3.1
accelerate==1.11.0
acres==0.5.0
aiohappyeyeballs==2.6.1
aiohttp==3.13.2
aiosignal==1.4.0
annexremote==1.6.6
annotated-types==0.7.0
ansi2html==1.9.2
anyio==4.8.0
arrow==1.3.0
asgiref==3.8.1
asrpy==0.0.3
astroid==4.0.2
astunparse==1.6.3
async-lru==2.0.5
async-timeout==5.0.1
autoflake==2.3.1
autoflake8==0.4.1
autopep8==2.3.2
babel==2.17.0
base==0.0.0
BCI2kReader==0.32.dev0
beautifulsoup4==4.13.4
bids-validator==1.14.7.post0
bids-validator-deno==2.1.0
bidsschematools==1.1.0
bitsandbytes==0.48.2
bleach==6.2.0
boto3==1.38.26
botocore==1.38.26
cachetools==5.5.0
certifi==2024.12.14
chardet==5.2.0
charset-normalizer==3.4.1
colorama==0.4.6
comm==0.2.2
crontab==1.0.1
cryptography==45.0.3
cut-cross-entropy==25.1.1
Cython==3.1.2
datalad==1.2.0
datasets==4.3.0
deep-translator==1.11.4
deepspeed==0.18.2
Deprecated==1.2.15
diffusers==0.35.2
distro==1.9.0
docopt==0.6.2
docstring_parser==0.17.0
dotenv==0.9.9
einops==0.8.1
et_xmlfile==2.0.0
executing==2.2.0
fastapi==0.110.0
fasteners==0.19
filelock==3.18.0
flatbuffers==25.2.10
fonttools==4.58.0
formulaic==1.2.1
freezegun==1.5.1
frozendict==2.4.7
frozenlist==1.8.0
fsspec==2025.5.1
ftfy==6.3.1
gast==0.6.0
geographiclib==2.0
geopy==2.4.1
google-api-core==2.22.0
google-api-python-client==2.151.0
google-auth==2.36.0
google-auth-httplib2==0.2.0
google-pasta==0.2.0
googleapis-common-protos==1.66.0
GraphRicciCurvature==0.5.3.2
greenlet==3.1.1
grpcio==1.62.3
h11==0.14.0
h2==4.2.0
h5io==0.2.5
h5py==3.13.0
hf-xet==1.1.5
hf_transfer==0.1.9
hjson==3.1.0
httpcore==1.0.6
httplib2==0.22.0
httpx==0.27.2
huggingface-hub==0.34.3
humanize==4.12.3
idna==3.10
imageio==2.37.0
importlib_metadata==8.6.1
iniconfig==2.1.0
interface-meta==1.3.0
ipyevents==2.0.2
ipympl==0.9.7
ipywidgets==8.1.7
iso8601==2.1.0
isort==7.0.0
jaraco.classes==3.4.0
jaraco.context==6.0.1
jaraco.functools==4.1.0
jeepney==0.9.0
jiter==0.8.2
jmespath==1.0.1
joblib==1.5.1
json5==0.12.0
jsonfixer==0.2.2
jsonschema==4.24.0
jupyter==1.1.1
jupyter-console==6.6.3
jupyter-lsp==2.2.5
jupyter_core==5.8.1
jupyterlab==4.4.3
jupyterlab_server==2.27.3
jupyterlab_widgets==3.0.15
keras==3.10.0
keyring==25.6.0
keyrings.alt==5.0.2
lazy_loader==0.4
libclang==18.1.1
lightning-utilities==0.15.2
llvmlite==0.44.0
lmdb==1.7.2
logfire==3.9.0
loguru==0.7.2
looseversion==1.3.0
lxml==5.4.0
Markdown==3.8.2
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.10.8
matplotlib-inline==0.1.7
mccabe==0.7.0
mistune==3.1.3
ml_dtypes==0.5.1
mne==1.9.0
mne-bids==0.16.0
mne-nirs==0.7.1
more-itertools==10.7.0
mpmath==1.3.0
msgpack==1.1.0
msgspec==0.19.0
multidict==6.7.0
multiprocess==0.70.16
munkres==1.1.4
namex==0.1.0
narwhals==2.13.0
networkit==11.1.post1
networkx==3.4.1
nilearn==0.11.1
ninja==1.13.0
nltk==3.9.1
notebook==7.4.3
num2words==0.5.14
numba==0.61.2
nvidia-cublas-cu12==12.8.4.1
nvidia-cuda-cupti-cu12==12.8.90
nvidia-cuda-nvcc-cu12==12.5.82
nvidia-cuda-nvrtc-cu12==12.8.93
nvidia-cuda-runtime-cu12==12.8.90
nvidia-cudnn-cu12==9.10.2.21
nvidia-cufft-cu12==11.3.3.83
nvidia-cufile-cu12==1.13.1.3
nvidia-curand-cu12==10.3.9.90
nvidia-cusolver-cu12==11.7.3.90
nvidia-cusparse-cu12==12.5.8.93
nvidia-cusparselt-cu12==0.7.1
nvidia-nccl-cu12==2.27.5
nvidia-nvjitlink-cu12==12.8.93
nvidia-nvshmem-cu12==3.3.20
nvidia-nvtx-cu12==12.8.90
openai==1.52.2
opencv-python==4.10.0.84
openpyxl==3.1.5
opentelemetry-api==1.31.1
opentelemetry-exporter-otlp-proto-common==1.31.1
opentelemetry-exporter-otlp-proto-http==1.31.1
opentelemetry-instrumentation==0.52b1
opentelemetry-instrumentation-asgi==0.52b1
opentelemetry-instrumentation-fastapi==0.52b1
opentelemetry-instrumentation-httpx==0.52b1
opentelemetry-instrumentation-sqlalchemy==0.52b1
opentelemetry-proto==1.31.1
opentelemetry-sdk==1.31.1
opentelemetry-semantic-conventions==0.52b1
opentelemetry-util-http==0.52b1
openTSNE==1.0.2
opt_einsum==3.4.0
optree==0.16.0
packaging==24.2
pandas==2.2.3
pathlib_abc==0.5.2
patool==4.0.1
patsy==1.0.1
pbd==0.9
peft==0.18.0
phonenumbers==8.13.55
pinecone==5.4.2
pinecone-client==6.0.0
pinecone-plugin-inference==3.1.0
pinecone-plugin-interface==0.0.7
pluggy==1.6.0
pooch==1.8.2
POT==0.9.5
propcache==0.4.1
proto-plus==1.25.0
protobuf==5.29.3
protoc-gen-openapiv2==0.0.1
psutil==7.0.0
ptflops==0.7.4
py-cpuinfo==9.0.0
pyarrow==22.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.1
pybids==0.21.0
pycodestyle==2.13.0
pycountry==24.6.1
pydantic==2.10.2
pydantic-settings==2.6.1
pydantic_core==2.27.1
pyEDFlib==0.1.42
pyflakes==3.4.0
Pygments==2.19.1
pykalman==0.10.1
pylint==4.0.4
PyMySQL==1.1.1
pynndescent==0.5.13
pyparsing==3.2.1
#PyQt5==5.15.11
#PyQt5-Qt5==5.15.17
#PyQt5_sip==12.17.0
pytest==8.4.0
python-dotenv==1.0.1
python-gitlab==6.0.0
python-louvain==0.16
pyvista==0.39.1
pyvistaqt==0.11.0
PyYAML==6.0.2
pyzmq==26.4.0
# QtPy==2.4.1
redis==4.3.4
referencing==0.36.2
regex==2024.11.6
requests==2.32.3
requests-toolbelt==1.0.0
rich==13.9.4
rq==2.1.0
rq-scheduler==0.14.0
rsa==4.9
ruff==0.12.12
s3transfer==0.13.0
safetensors==0.5.3
scikit-base==0.12.3
scikit-learn==1.6.1
scooby==0.10.1
seaborn==0.13.2
SecretStorage==3.3.3
segment-anything==1.0
sentencepiece==0.2.1
setuptools==80.9.0
shtab==1.7.2
snakeviz==2.2.2
snirf==0.8.0
soupsieve==2.7
SQLAlchemy==2.0.36
sqlmodel==0.0.22
starlette==0.36.3
statsmodels==0.14.4
stripe==11.4.1
sympy==1.14.0
tabulate==0.9.0
templateflow==25.1.1
tensorboard==2.20.0
tensorboard-data-server==0.7.2
tensorflow==2.20.0
termcolor==3.1.0
tf_keras==2.20.1
threadpoolctl==3.6.0
tinycss2==1.4.0
tokenizers==0.22.1
tomlkit==0.13.3
torch==2.9.0
torchao==0.14.1
torchinfo==1.8.0
torchmetrics==1.8.1
torchvision==0.24.0
tornado==6.5.1
tqdm==4.67.1
transformers==4.57.1
triton==3.5.0
# trl==0.8.6
typeguard==4.4.4
types-python-dateutil==2.9.0.20250516
tyro==0.9.35
tzdata==2025.1
umap-learn==0.5.9.post2
universal_pathlib==0.3.6
unsloth==2025.11.3
unsloth_zoo==2025.11.4
uritemplate==4.1.1
urllib3==2.3.0
uvicorn==0.30.0
# vtk==9.3.0
waiting==1.5.0
webcolors==24.11.1
websocket-client==1.8.0
Werkzeug==3.1.3
wheel==0.45.1
widgetsnbextension==4.0.14
win32_setctime==1.2.0
wrapt==1.17.2
# wxPython==4.2.4
xformers==0.0.33.post1
xgboost==2.1.1
xlrd==2.0.1
XlsxWriter==3.2.3
xxhash==3.6.0
yarl==1.22.0
zipp==3.21.0
zstandard==0.25.0
